Caution.md
Although this is good example for a blog website, however, there are a few errors, e.g LIKE icon counter is not functional and I'm unable to find a correct solution, so please take care of it, if you can correct the error in code, please make a branch of the repo. Rest everything is smooth and correct.